Fixed several problems with configure.ac.
Updated the version to beta3.

Removed the use of getting 'pg_config --pkgincludedir' since it is not
needed to build ptop.

Removed the explicit setting of the RPATH for the linker, as RPM
building warns against it.  There are other good reasons for it too,
relating letting the shared libraries be found in the environment path.