Updates for openbsd
Did some header clean-up and isolating of more variables, the below
patch is required to make it compile and give something back in
-current.  In addition, there is a change to get rid of stathz to make
the source of process TIME for displaying and sorting the same - this
matches the base top(1) and ps(1) - made it easier for me to see what
each tool was doing, so I'm including the fix in the diff below.

However, this doesn't seem to fix everything, and I think it has to do
with OpenBSD's threads (rthreads).